renamed new solver-job service
