+++ services/api-server/src/simcore_service_api_server/api/routes/solvers_jobs.py
@@ -32,13 +32,13 @@ from ...models.schemas.jobs import (
 from ...models.schemas.solvers import Solver, SolverKeyId
 from ...services.catalog import CatalogApi
 from ...services.director_v2 import DirectorV2Api, DownloadLink, NodeName
-from ...services.storage import StorageApi, to_file_api_model
-from ...utils.solver_job_models_converters import (
+from ...services.solver_job_models_converters import (
     create_job_from_project,
     create_jobstatus_from_task,
     create_new_project_for_job,
 )
-from ...utils.solver_job_outputs import ResultsTypes, get_solver_output_results
+from ...services.solver_job_outputs import ResultsTypes, get_solver_output_results
+from ...services.storage import StorageApi, to_file_api_model
 from ..dependencies.application import get_product_name, get_reverse_url_mapper
 from ..dependencies.authentication import get_current_user_id
 from ..dependencies.database import Engine, get_db_engine
