+++ services/api-server/tests/unit/test_services_solver_job_models_converters.py
@@ -9,7 +9,7 @@ from pydantic import create_model, parse_obj_as
 from simcore_service_api_server.models.schemas.files import File
 from simcore_service_api_server.models.schemas.jobs import ArgumentTypes, Job, JobInputs
 from simcore_service_api_server.models.schemas.solvers import Solver
-from simcore_service_api_server.utils.solver_job_models_converters import (
+from simcore_service_api_server.services.solver_job_models_converters import (
     create_job_from_project,
     create_job_inputs_from_node_inputs,
     create_jobstatus_from_task,
@@ -108,7 +108,6 @@ def test_job_to_node_inputs_conversion():
 
 
 def test_create_job_from_project():
-
     project = Project.parse_obj(
         {
             "uuid": "f925e30f-19de-42dc-acab-3ce93ea0a0a7",
